Include demographic information in analytics event
Since the demographic custom Google Analytics dimension is set with the user scope, Analytics will remember it and apply it to this same user's events in the future if it is missing from subsequent events. This commit avoids sending a NULL value for this commit when the user doesn't select a demographic group, which would overwrite the previous non-null value and prevent Analytics from effectively using it in the user scope.