Make year restriction on statistics be per-metric
...instead of being a single year for all metrics. Before this, metrics
with more recent stats were forcing metrics with older stats to
contribute only null values to the ranking formula. I realized that the
spirit of the "restrict it to one year" rule made no sense with one year
applying to all metrics, and that one year per metric was appropriate.