Move Elasticsearch index name to env var
Instead of the index name always being 'statistics', it's now going to
reflect the time and parameters of its generation, e.g.
"statistics-nohidden-recent-2020-01-08". Since this value will change
whenever statistical data is updated and is kind of a separate matter
from the codebase, it made more sense to move this out of committed
files and into environment variables.